+++ modules/relax/includes/js/views/widgets/relaxer.js
@@ -1,0 +1,21 @@
+/*! Copyright 2016 - Ortus Solutions (Compiled: 18-03-2016) */
+define([ "Backbone" ], function(Backbone) {
+    "use strict";
+    var View = Backbone.View.extend({
+        el: "#main-content",
+        events: {},
+        initialize: function() {
+            return this.setupDefaults().setupSelectors().render();
+        },
+        setupSelectors: function() {
+            return this;
+        },
+        setupDefaults: function() {
+            return this;
+        },
+        render: function() {
+            return this;
+        }
+    });
+    return new View();
+});